Modernize resume build and dev scripts
This brings us into the modern days of React and associated Webpack
nonsense for the resume. A big benefit of this is smaller builds: the JS
for the page clocks in at around 150kb. Additionally, now that Webpack
can consume javascript modules natively, we cut down on the filesize and
performance problems --- there's less of a cost to wrapping up many
small modules now that they're inlined rather than going through some
function scope. Other goodies:

- Yarn for locking dependencies
- Removal of separate webpack configs for dev and prod, in favor of one
that augments the config if imported into the server.js file (the dev
server) or if `process.env` is set to production for cli builds
- No more `/static/` base in dev, and `dist/bundle.js` file in prod.