Build initial markup from the react component
Not a super clean PR, but a useful one: create two builds for the
project, one where server rendering and hot-reloading is enabled, and
one where we render the intial HTML for the resume and stick that into
the index.html file picked up by Jekyll for initial rendering.

To get this to work, extracting the css into its own file rather than
going through webpack's loader was essential. This also seemed like
a good opportunity to upgrade from the older autoprefixer-loader to the
new postcss-loader flavor. It wasn't a good opportunity though.