Move Jekyll site source into a subdirectory
Because this project has generated files, it would be best to not check
them into source control. Just like we ignore the output of the Jekyll
build directory (_site), we should also ignore the `bundle.js` files
generated from the homepage rollup script and from the resume build
script.

This is the first step in that direction --- it promotes the
package.json in the js directory to be top-level, and moves all the
Jekyll source material into the src directory. It then builds the
homepage js into the _site/js directory, so that it will be copied over
when publishing the site, but doesn't have to be checked into version
control.

The next step will be merging the build contained in the src/resume
directory into this top-level scheme, perhaps making a better source
javascript directory and consolidating the builds so that both files go
through webpack.