Move resume outside of jekyll source
This allows for the benefits mentioned in the previous commit, where we
can stop committing the build. Not as painful a change as I thought it
would be. There's probably a lot that can be cleaned up still,
especially around generating the intiial HTML for the resume and
sticking that file into the source directory, but for now, we're at
parity with the previous version.