Upgrade rollup build deps
Rollup is only used for the checkboxes feature right now, but I have
a feeling I'm going to upgrade the webpack-related toolchain to use
it. This updates to Rollup 1.x, and bumps all rollup plugins, except for
the svelte one. New version of svelte has a bunch of breaking changes,
so I'm going to do that separately.